ch(fastfoodfast): split views file into smaller files
Put related views in a given file

[Delivers #160982299]
