+++ config.py
@@ -1,6 +1,2 @@
 """ Config file to handle app configurations """
-from environs import Env
-
 DEBUG = True
-env = Env()
-env.read_env()
